rename AvroSerializationError to AvroSerializerError
