+++ sdk/schemaregistry/schema-registry-avro/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.0.0-beta.8 (Unreleased)
 
 ### Features Added
-- A new error type, `AvroSerializationError`, is added and is thrown in all error cases except for service calls.
+- A new error type, `AvroSerializerError`, is added and is thrown in all error cases except for service calls.
 
 ### Breaking Changes
