+++ sdk/schemaregistry/schema-registry-avro/review/schema-registry-avro.api.md
@@ -7,20 +7,20 @@
 import { SchemaRegistry } from '@azure/schema-registry';
 
 // @public
-export class AvroSerializationError extends Error {
+export class AvroSerializer<MessageT = MessageWithMetadata> {
+    constructor(client: SchemaRegistry, options?: AvroSerializerOptions<MessageT>);
+    deserializeMessageData(message: MessageT, options?: DeserializeMessageDataOptions): Promise<unknown>;
+    serializeMessageData(value: unknown, schema: string): Promise<MessageT>;
+}
+
+// @public
+export class AvroSerializerError extends Error {
     constructor(message: string, options?: {
         innerError?: unknown;
         schemaId?: string;
     });
     innerError?: unknown;
     schemaId?: string;
-}
-
-// @public
-export class AvroSerializer<MessageT = MessageWithMetadata> {
-    constructor(client: SchemaRegistry, options?: AvroSerializerOptions<MessageT>);
-    deserializeMessageData(message: MessageT, options?: DeserializeMessageDataOptions): Promise<unknown>;
-    serializeMessageData(value: unknown, schema: string): Promise<MessageT>;
 }
 
 // @public
