+++ sdk/schemaregistry/schema-registry-avro/src/avroSerializer.ts
@@ -3,7 +3,7 @@
 
 import * as avro from "avsc";
 import {
-  AvroSerializationError,
+  AvroSerializerError,
   AvroSerializerOptions,
   DeserializeMessageDataOptions,
   MessageAdapter,
@@ -70,7 +70,7 @@ export class AvroSerializer<MessageT = MessageWithMetadata> {
    * @param schema - The Avro schema to use.
    * @returns A new message with the serialized value. The structure of message is
    * constrolled by the message factory option.
-   * @throws {@link AvroSerializationError}
+   * @throws {@link AvroSerializerError}
    * Thrown if the schema can not be parsed or the value does not match the schema.
    */
   async serializeMessageData(value: unknown, schema: string): Promise<MessageT> {
@@ -111,7 +111,7 @@ export class AvroSerializer<MessageT = MessageWithMetadata> {
    * @param message - The message with the payload to be deserialized.
    * @param options - Decoding options.
    * @returns The deserialized value.
-   * @throws {@link AvroSerializationError}
+   * @throws {@link AvroSerializerError}
    * Thrown if the deserialization failed, e.g. because reader and writer schemas are incompatible.
    */
   async deserializeMessageData(
@@ -160,11 +160,11 @@ export class AvroSerializer<MessageT = MessageWithMetadata> {
 
     const schemaResponse = await this.registry.getSchema(schemaId);
     if (!schemaResponse) {
-      throw new AvroSerializationError(`Schema with ID '${schemaId}' not found.`);
+      throw new AvroSerializerError(`Schema with ID '${schemaId}' not found.`);
     }
 
     if (!schemaResponse.properties.format.match(/^avro$/i)) {
-      throw new AvroSerializationError(
+      throw new AvroSerializerError(
         `Schema with ID '${schemaResponse.properties.id}' has format '${schemaResponse.properties.format}', not 'avro'.`
       );
     }
@@ -183,11 +183,11 @@ export class AvroSerializer<MessageT = MessageWithMetadata> {
 
     const avroType = getSerializerForSchema(schema);
     if (!avroType.name) {
-      throw new AvroSerializationError("Schema must have a name.");
+      throw new AvroSerializerError("Schema must have a name.");
     }
 
     if (!this.schemaGroup) {
-      throw new AvroSerializationError(
+      throw new AvroSerializerError(
         "Schema group must have been specified in the constructor options when the client was created in order to serialize."
       );
     }
@@ -207,7 +207,7 @@ export class AvroSerializer<MessageT = MessageWithMetadata> {
         id = (await this.registry.getSchemaProperties(description)).id;
       } catch (e) {
         if (e.statusCode === 404) {
-          throw new AvroSerializationError(
+          throw new AvroSerializerError(
             `Schema '${description.name}' not found in registry group '${description.groupName}', or not found to have matching definition.`
           );
         } else {
@@ -233,12 +233,12 @@ export class AvroSerializer<MessageT = MessageWithMetadata> {
 function getSchemaId(contentType: string): string {
   const contentTypeParts = contentType.split("+");
   if (contentTypeParts.length !== 2) {
-    throw new AvroSerializationError(
+    throw new AvroSerializerError(
       "Content type was not in the expected format of MIME type + schema ID"
     );
   }
   if (contentTypeParts[0] !== avroMimeType) {
-    throw new AvroSerializationError(
+    throw new AvroSerializerError(
       `Received content of type ${contentTypeParts[0]} but an avro serializer may only be used on content that is of '${avroMimeType}' type`
     );
   }
@@ -274,7 +274,7 @@ function convertMessage<MessageT>(
   } else if (isMessageWithMetadata(message)) {
     return convertPayload(message.body, message.contentType);
   } else {
-    throw new AvroSerializationError(
+    throw new AvroSerializerError(
       `Expected either a message adapter to be provided to the serializer or the input message to have body and contentType fields`
     );
   }
@@ -331,7 +331,7 @@ function wrapError<T>(
     result = f();
   } catch (innerError) {
     const { schemaId } = options;
-    throw new AvroSerializationError(message, {
+    throw new AvroSerializerError(message, {
       innerError,
       schemaId,
     });
